Add pyproject.toml and setup.cfg for 'python -m pip ...' commands.
setup.py retained for 'python setup.py ...' commands until this is no
longer supported, but modified to rely on setup.cfg.

Bump version number with .dev0 suffix to test providing a URL, or path
to a html file, containing links to archives for the --find-links
argument to 'python -m pip ...' commands.